+++ litebird_sim/hwp_sys/hwp_sys.py
@@ -7,13 +7,14 @@ import numpy as np
 from astropy import constants as const
 from astropy.cosmology import Planck18 as cosmo
 from numba import njit, prange
+
 import litebird_sim as lbs
 from litebird_sim import mpi
 from ..coordinates import rotate_coordinates_e2g
 from ..detectors import FreqChannelInfo
+from ..hwp_diff_emiss import compute_2f_for_one_sample
 from ..mbs.mbs import MbsParameters
 from ..observations import Observation
-from ..hwp_diff_emiss import compute_2f_for_one_sample
 
 COND_THRESHOLD = 1e10
 
@@ -272,13 +273,14 @@ def compute_signal_for_one_sample(
 @njit(parallel=True)
 def compute_signal_for_one_detector(
     tod_det,
-    pixel_ind,
     m0f,
     m2f,
     m4f,
     rho,
     psi,
-    maps,
+    mapT,
+    mapQ,
+    mapU,
     cos2Xi2Phi,
     sin2Xi2Phi,
     phi,
@@ -296,9 +298,9 @@ def compute_signal_for_one_detector(
         Four_rho_phi = 4 * (rho[i] - phi)
         Two_rho_phi = 2 * (rho[i] - phi)
         tod_det[i] += compute_signal_for_one_sample(
-            T=maps[0, pixel_ind[i]],
-            Q=maps[1, pixel_ind[i]],
-            U=maps[2, pixel_ind[i]],
+            T=mapT[i],
+            Q=mapQ[i],
+            U=mapU[i],
             mII=m0f[0, 0]
             + m2f[0, 0] * np.cos(Two_rho_phi - 2.32)
             + m4f[0, 0] * np.cos(Four_rho_phi - 0.84),
@@ -459,6 +461,7 @@ class HwpSys:
         build_map_on_the_fly: Union[bool, None] = False,
         apply_non_linearity: Union[bool, None] = False,
         add_2f_hwpss: Union[bool, None] = False,
+        interpolation: Union[str, None] = "",
         Channel: Union[FreqChannelInfo, None] = None,
         maps: Union[np.ndarray, None] = None,
         comm: Union[bool, None] = None,
@@ -473,6 +476,9 @@ class HwpSys:
           apply_non_linearity (bool): applies the coupling of the non-linearity
               systematics with hwp_sys
           add_2f_hwpss (bool): adds the 2f hwpss signal to the TOD
+          interpolation (str): if it is ``""`` (the default), pixels in the map
+              won’t be interpolated. If it is ``linear``, a linear interpolation
+              will be used
           Channel (:class:`.FreqChannelInfo`): an instance of the
                                                 :class:`.FreqChannelInfo` class
           maps (float): input maps (3, npix) coherent with nside provided,
@@ -560,6 +566,8 @@ class HwpSys:
             Mbsparams.nside = self.nside
 
         self.npix = hp.nside2npix(self.nside)
+
+        self.interpolation = interpolation
 
         if Channel is None:
             Channel = lbs.FreqChannelInfo(bandcenter_ghz=140)
@@ -768,7 +776,37 @@ class HwpSys:
 
                 # all observed pixels over time (for each sample),
                 # i.e. len(pix)==len(times)
-                pix = hp.ang2pix(self.nside, cur_point[:, 0], cur_point[:, 1])
+                if self.interpolation in ["", None] or self.build_map_on_the_fly:
+                    pix = hp.ang2pix(self.nside, cur_point[:, 0], cur_point[:, 1])
+
+                if self.interpolation in ["", None]:
+                    input_T = self.maps[0, pix]
+                    input_Q = self.maps[1, pix]
+                    input_U = self.maps[2, pix]
+
+                elif self.interpolation == "linear":
+                    input_T = hp.get_interp_val(
+                        self.maps[0, :],
+                        cur_point[:, 0],
+                        cur_point[:, 1],
+                    )
+                    input_Q = hp.get_interp_val(
+                        self.maps[1, :],
+                        cur_point[:, 0],
+                        cur_point[:, 1],
+                    )
+                    input_U = hp.get_interp_val(
+                        self.maps[2, :],
+                        cur_point[:, 0],
+                        cur_point[:, 1],
+                    )
+                else:
+                    raise ValueError(
+                        "Wrong value for interpolation. It should be one of the following:\n"
+                        + '- "" for no interpolation\n'
+                        + '- "linear" for linear interpolation\n'
+                    )
+
                 # separating polarization angle xi from cur_point[:, 2] = psi + xi
                 # xi: polarization angle, i.e. detector dependent
                 # psi: instrument angle, i.e. boresight direction from focal plane POV
@@ -782,13 +820,14 @@ class HwpSys:
 
                 compute_signal_for_one_detector(
                     tod_det=tod,
-                    pixel_ind=pix,
                     m0f=cur_obs.mueller_hwp[idet]["0f"],
                     m2f=cur_obs.mueller_hwp[idet]["2f"],
                     m4f=cur_obs.mueller_hwp[idet]["4f"],
                     rho=np.array(cur_hwp_angle, dtype=np.float64),
                     psi=np.array(psi, dtype=np.float64),
-                    maps=self.maps,
+                    mapT=input_T,
+                    mapQ=input_Q,
+                    mapU=input_U,
                     cos2Xi2Phi=cos2Xi2Phi,
                     sin2Xi2Phi=sin2Xi2Phi,
                     phi=phi,
@@ -816,7 +855,9 @@ class HwpSys:
 
                 cur_obs.tod[idet] = tod
 
-        del pix
+        if self.interpolation in ["", None]:
+            del pix
+        del input_T, input_Q, input_U
         if not save_tod:
             del cur_obs.tod